Refactor utilities code in preparation to move to external repo
