+++ test/tests/utilitiesTest.js
@@ -508,51 +508,6 @@ describe("Zotero.Utilities", function() {
 		})
 	});
 	
-	describe("#parseURL()", function () {
-		var f;
-		before(() => {
-			f = Zotero.Utilities.parseURL;
-		});
-		
-		describe("#fileName", function () {
-			it("should contain filename", function () {
-				assert.propertyVal(f('http://example.com/abc/def.html?foo=bar'), 'fileName', 'def.html');
-			});
-			
-			it("should be empty if no filename", function () {
-				assert.propertyVal(f('http://example.com/abc/'), 'fileName', '');
-			});
-		});
-		
-		describe("#fileExtension", function () {
-			it("should contain extension", function () {
-				assert.propertyVal(f('http://example.com/abc/def.html?foo=bar'), 'fileExtension', 'html');
-			});
-			
-			it("should be empty if no extension", function () {
-				assert.propertyVal(f('http://example.com/abc/def'), 'fileExtension', '');
-			});
-			
-			it("should be empty if no filename", function () {
-				assert.propertyVal(f('http://example.com/abc/'), 'fileExtension', '');
-			});
-		});
-		
-		describe("#fileBaseName", function () {
-			it("should contain base name", function () {
-				assert.propertyVal(f('http://example.com/abc/def.html?foo=bar'), 'fileBaseName', 'def');
-			});
-			
-			it("should equal filename if no extension", function () {
-				assert.propertyVal(f('http://example.com/abc/def'), 'fileBaseName', 'def');
-			});
-			
-			it("should be empty if no filename", function () {
-				assert.propertyVal(f('http://example.com/abc/'), 'fileBaseName', '');
-			});
-		});
-	});
-	
 	
 	describe("#ellipsize()", function () {
 		describe("with wordBoundary", function () {
